fix: 빌드 에러 해결

See merge request s11-blochain-transaction-sub1/S11P21C107!224

+++ backend/main-service/src/main/java/com/turtlecoin/mainservice/domain/transaction/dto/DetailTransactionResponseDto.java
@@ -15,7 +15,7 @@ public class DetailTransactionResponseDto {
     private String scientificName;
     private Double price;
     private String createDate;
-    private Double weight;
+    private int weight;
     private String content;
     private List<String> transactionTag;  // 태그 목록
     private List<String> transactionImage;  // 이미지 경로 목록
